Infinite asset query refactor
