+++ client/packages/coldchain/src/Equipment/ListView/CreateAssetModal.tsx
@@ -24,6 +24,7 @@ import {
   StoreSearchInput,
   mapIdNameToOptions,
   useAssetData,
+  useInfiniteAssets,
 } from '@openmsupply-client/system';
 import { useAssets } from '../api';
 import { CCE_CLASS_ID } from '../utils';
@@ -96,8 +97,12 @@ export const CreateAssetModal = ({
     data: catalogueItemData,
     isFetching,
     fetchNextPage,
-  } = useAssetData.document.infiniteList({
-    filter,
+  } = useInfiniteAssets({
+    queryParams: {
+      filterBy: {
+        ...filter,
+      },
+    },
     categoryId: draft.categoryId,
     rowsPerPage: RECORDS_PER_PAGE,
   });
