+++ client/packages/system/src/Asset/api/hooks/document/index.ts
@@ -1,11 +1,7 @@
 import { useAsset } from './useAsset';
-import { useAssetFields } from './useAssetFields';
-import { useInfiniteAssets } from './useInfiniteAssets';
 import { useAssetsDelete } from './useAssetsDelete';
 
 export const Document = {
   useAsset,
-  useAssetFields,
-  useInfiniteAssets,
   useAssetsDelete,
 };
