--- client/packages/system/src/Asset/api/hooks/document/useAssetFields.ts
@@ -1,6 +1,0 @@
-import { useAsset } from './useAsset';
-
-export const useAssetFields = () => {
-  const { data } = useAsset();
-  return { ...data };
-};
