+++ client/packages/system/src/Asset/api/hooks/index.ts
@@ -14,9 +14,7 @@ export const useAssetData = {
 
   document: {
     get: Document.useAsset,
-    infiniteList: Document.useInfiniteAssets,
 
-    fields: Document.useAssetFields,
     delete: Document.useAssetsDelete,
   },
