+++ client/packages/system/src/Asset/index.ts
@@ -4,5 +4,6 @@ export {
   AssetPropertyFragment,
   useAssetData,
   useAssetList,
+  useInfiniteAssets,
 } from './api';
 export { mapIdNameToOptions } from './utils';
